show alets on confirmation add, delete and await

+++ src/modules/daybook/views/EntryView.vue
@@ -40,6 +40,7 @@ import { defineAsyncComponent } from 'vue';
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'; // computed!!!
 import getDayMonthYear from '@/modules/daybook/helpers/getDayMonthYear'
+import Swal from 'sweetalert2'
 
 export default {
     props: {
@@ -92,6 +93,12 @@ export default {
             this.entry = entry
         },
         async saveEntry() {
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
             if( this.entry.id ) {
                 // update entry
                 await this.updateEntries( this.entry )
@@ -99,12 +106,30 @@ export default {
                 // create new entry
                 const id = await this.createEntry( this.entry )
                 this.$router.push({ name: 'entry', params: { id }})
-            }  
+            }
+
+            Swal.fire( 'Guardado', 'Entrada registrada con éxito', 'success' )
         },
         async onDeleteEntry() {
-            this.deleteEntry( this.entry.id )
-            // redirect user
-            this.$router.push({ name: 'no-entry' })
+
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Sí, estoy seguro'
+            })
+
+            if( isConfirmed === true ) {
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false,
+                })
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )
+                // redirect user
+                this.$router.push({ name: 'no-entry' })
+                Swal.fire( 'Eliminado', '', 'success' )
+            }            
         }
     },
     created() {
